added some authentication for delete so user owner only deletes it

+++ app/routers/Posts.py
@@ -24,9 +24,10 @@ def get_latest_post(db: Session=Depends(get_db)):
 
 #POST BY ID
 @router.get("/{id}",response_model=PostResponse)
-def get_posts(id: int,db: Session=Depends(get_db),user_id=Depends(oauth2.get_user)):
+def get_posts(id: int,db: Session=Depends(get_db),user=Depends(oauth2.get_user)):
     # cursor.execute("select * from posts where id=%s",(id,))
     # post_by_id=cursor.fetchone()
+    print(user.email)
     post_by_id=db.query(models.Post).filter(models.Post.id==id).first()
     if post_by_id:
         return post_by_id
@@ -34,12 +35,12 @@ def get_posts(id: int,db: Session=Depends(get_db),user_id=Depends(oauth2.get_use
    
 #CREATE POST
 @router.post("/",status_code=status.HTTP_201_CREATED,response_model=PostResponse)
-def create_post(payload: PostCreate,db: Session=Depends(get_db),user_id=Depends(oauth2.get_user)):
+def create_post(payload: PostCreate,db: Session=Depends(get_db),current_user=Depends(oauth2.get_user)):
     # cursor.execute("""INSERT INTO posts(title,content,is_published) VALUES (%s,%s,%s) RETURNING *""",(payload.title,payload.content,payload.published))
     # new_post=cursor.fetchone()
     # conn.commit()
-    print(user_id)
-    new_post = models.Post(**payload.model_dump())   
+    
+    new_post = models.Post(**payload.model_dump(),user_id=current_user.id)   
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -47,14 +48,16 @@ def create_post(payload: PostCreate,db: Session=Depends(get_db),user_id=Depends(
 
 
 #DELETE
-@router.delete('/{id}')
-def delete_post(id: int,db: Session=Depends(get_db),user_id=Depends(oauth2.get_user)):
+@router.delete('/{id}')  
+def delete_post(id: int,db: Session=Depends(get_db),current_user=Depends(oauth2.get_user)):
     # cursor.execute("""DELETE FROM posts where id=%s RETURNING *""",(id,))
     # post_by_id=cursor.fetchone()
     # conn.commit()
     post_by_id=db.query(models.Post).filter(models.Post.id==id).first()
 
     if post_by_id:
+        if post_by_id.user_id!=current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail=f"Not Authorised to DELETE THIS POST") 
         db.delete(post_by_id)
         db.commit()
         return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -62,7 +65,7 @@ def delete_post(id: int,db: Session=Depends(get_db),user_id=Depends(oauth2.get_u
 
 #UPDATE
 @router.put('/{id}',response_model=PostResponse)
-def update_post(id: int, upd_post:PostCreate,db: Session=Depends(get_db),user_id=Depends(oauth2.get_user)):
+def update_post(id: int, upd_post:PostCreate,db: Session=Depends(get_db),current_user=Depends(oauth2.get_user)):
     # cursor.execute("""UPDATE POSTS SET title=%s, content=%s, is_published=%s WHERE ID =%s RETURNING *""", (upd_post.title,upd_post.content,upd_post.published,id))
     # new_post=cursor.fetchone()
     # conn.commit()
@@ -70,6 +73,8 @@ def update_post(id: int, upd_post:PostCreate,db: Session=Depends(get_db),user_id
     post=post_query.first()
     
     if post:
+        if post.user_id!=current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail=f"Not Authorised to Update this Post")
         post_query.update(upd_post.model_dump(),synchronize_session=False)
         db.commit()
         db.refresh(post)
